LPS-146537 Rename the info-screen directory to InfoScreen to follow a pattern
